requirements/development.txt
-r base.txt

pytest
pytest-cov
pytest-django
pep8
rope
pylint
